fixing EWR logic test
changed the order of the if statements.

also added a column to describe time aggregation 'ewr_achieved_timeframe' maybe need to rename?

d

also, probably doesn't work properly if missing years.
also using the NSW method of using 10 years for some EWRs and 20 years for others. i dont think this is actually an appropriate thing to do. but should discuss with martin