GEOS & GDAL for ubuntu/mac os
Github actions testing is working for windows OS but not when running on ubuntu or mac os. The error log on github indicates the issue is with the test-spatial_aggregate.R file line 81 that uses expect_doppleganger(). This function requires successful rendering of ggplots using spatial data, according to the link below these will only work if system libraries geos and gdal are installed on the OS. 

See https://vdiffr.r-lib.org/reference/expect_doppelganger.html

To ensure geos and gdal are installed on ubuntu and mac os rnuners, I added installation steps into the R-CMD-check.yaml that conditionally install these system libraries if the os is Ubuntu or macOS.