CI check (bug edit & rsvg lib added)
Forgot to change conditional if OS==ubutnu.20.04 after editing the matrix.config.os. Added that.

I'm not sure this will work though because the versioning of ubuntu apparently isn't the issue but rather the rsvg library. Because of this I added rsvg library to install on ubuntu to see if helps. Mac OS issues still look related to gdal/geos. The error is "crs not found: is it missing?" which is Coordinate Reference System (CRS) should be there from gdal install, adjusted code to include | which runs multiple lines to see if helps.

If it doesn't work again consider skipping these tests on ubuntu/mac.