Check CI (edited ubuntu systems lib install)
The step installing ubuntu system libraries was using '-dev' suffix which is more suitable for development environments instead of use by interfacing programs like R so changed the libs to -bin or -lib or -2 after some research. These should be more appropriate and suited for just passing a run on github runner.