Skipping more tests
Adding more ocnditional if statement ot only run on windows for one more test containing expect_doppelganger() there's one more at least in this test-spatial_aggregate file. There may be more. unfortunately there is not built in continue-on-error argument for RMD checks but I could try and figure it out but difficult without having an interactive console to bash in with the back and forth of github actions so just doing it the messy way.