CI check - setting proj lib library
The errors failing mac runs are related to proj systems package. The error (below) indicates issues with finding the proj database so I added a line of code in the workflow yaml to set the path and see if that fixes the issue.

Error from github runner:

"GDAL Error 1: PROJ: proj_identify: Cannot find proj.db"