Making changes to .pre-commit-config.yaml to include a parsable test to make sure the code runs (excluding lintr hook for now as may be too much). Other available hooks like spell check can  be found here ->https://lorenzwalthert.github.io/precommit/articles/available-hooks.html
More info: I found that pre-commit hooks only run when running git commit through the R studio interface and throw errors in gitKraken or Gitbash. I think this may be because a commit doens't process unless all the checks in the pre-commit yaml pass, so it needs to be run twice effectively if it makes any changes. So committing in R works maybe because it connects to make changes more seemlessly when in gitKraken or gitbash it breaks and doesn't allow you to re-run/process the automated modifications.